backend/requirements.txt
# Backend Python dependencies
# Versions pinned for reproducibility (as of 2026-01-20)

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Configuration and validation
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# Authentication
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
email-validator==2.3.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
aiosqlite==0.20.0
httpx==0.28.1

# Additional uvicorn dependencies (via [standard])
# These are included automatically but listed for clarity:
# - uvloop (for faster async event loop)
# - httptools (for faster HTTP parsing)
# - websockets (for WebSocket support)
